# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel and Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "an.nguyenbku1999@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/project/add": {
            "post": {
                "tags": [
                    "Project"
                ],
                "description": "Add Project",
                "parameters": [
                    {
                        "name": "UserId",
                        "in": "query",
                        "description": "Enter your UserId:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter your project name:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "urlAvatar",
                        "in": "query",
                        "description": "Enter your urlAvatar",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/project/delete": {
            "post": {
                "tags": [
                    "Project"
                ],
                "description": "Delete Project",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your user token when logged in:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "ProjectId",
                        "in": "query",
                        "description": "Enter your project id:",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/project/edit": {
            "post": {
                "tags": [
                    "Project"
                ],
                "description": "Edit Project",
                "parameters": [
                    {
                        "name": "ProjectId",
                        "in": "query",
                        "description": "Enter your project name:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter your project name:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "urlAvatar",
                        "in": "query",
                        "description": "Enter your urlAvatar",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/project/addMembertoProject": {
            "post": {
                "tags": [
                    "Project"
                ],
                "description": "add Member to Project",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "UserIdList",
                        "in": "query",
                        "description": "Enter User Member List like [1,2,3,...]",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/project/removeMember": {
            "post": {
                "tags": [
                    "Project"
                ],
                "description": "remove Member from Project",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Enter key project get from showProjectsList api:",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "UserId",
                        "in": "query",
                        "description": "Enter User ID Member do you want to remove:",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
        "/api/project/showProjectsList": {
            "get": {
                "tags": [
                    "Project"
                ],
                "description": "show Projects List",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Enter your token when you logged in:",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful."
                    },
                    "204": {
                        "description": "No Content."
                    },
                    "405": {
                        "description": "Method Not Allowed."
                    },
                    "409": {
                        "description": "The request could not be completed due to a conflict with the current state of the target resource."
                    },
                    "422": {
                        "description": "Missing Data or Data is incorrect."
                    }
                }
            }
        },
...
    },
...
}